Replace tasks array with dictionary
The dictionary looks better than an array filled with nils which was a
solution imported from the idea with observables (@State or @Published)
but unnecessary for a regular property. Also a dictionary allows us to
reuse tasks for images that happen to have the same resourceURI although
that may be unlikely to happen.
In order to make the dict easier to implement, I've also prefiltered for
non-nil resourceURIs, removing useless rows from display.